requirements.txt
scipy==1.1.0
resampy==0.2.1
tensorflow-gpu==1.8.0
six==1.11.0
pandas==0.23.4
keras==2.2.4
scikit-learn==0.20.0